+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Add `EnvelopeProperties.auditLogAvailableToAllRecipients`
 - Add `EnvelopeRecipient.authFailedReason`
 - Update `EnvelopesEndpoint.discard` with body like discardAt param
+- Add new `EnvelopeDocumentSignatureSheets` resource and endpoint
 
 ## [1.10.0]
 ### Added